# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Plastic Hunter AI API
  version: 1.0.0
  description: Runtime API for the Plastic Hunter AI FastAPI demo.
servers:
  - url: /
paths:
...
  /detect:
    post:
      summary: Upload an image and store a lightweight CV demo detection
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                latitude:
                  type: number
                  minimum: -90
                  maximum: 90
                longitude:
                  type: number
                  minimum: -180
                  maximum: 180
      responses:
        "200":
          description: Detection result
        "400":
          description: Invalid or empty image upload
        "413":
          description: Image exceeds upload limit
        "422":
          description: Invalid image or coordinate values
...
  /results/{filename}:
    get:
      summary: Serve an annotated result image
      parameters:
        - in: path
          name: filename
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Annotated JPEG image
        "400":
          description: Invalid filename
        "404":
          description: Image not found
...